Task/requirements.txt
django>=3.2
djangorestframework>=3.12
pymongo>=4.0
djongo>=1.3
PyJWT>=2.0
